==> earlier code/5. get ready to generate features.ipynb <==
import pandas as pd

train_mkt_data = pd.read_csv("train_mkt_data.csv",header =0)
train_mkt_data = train_mkt_data.drop("Unnamed: 0",axis=1)
train_mkt_data.head(5)

valid_mkt_data = pd.read_csv("valid_mkt_data.csv",header =0)
valid_mkt_data = valid_mkt_data.drop("Unnamed: 0",axis=1)
valid_mkt_data.head(5)

train_ind_data = pd.read_csv("train_ind_data.csv",header = 0)
train_ind_data = train_ind_data.drop("Unnamed: 0",axis=1)
train_ind_data.head(5)

valid_ind_data = pd.read_csv("valid_ind_data.csv",header = 0)
valid_ind_data = valid_ind_data.drop("Unnamed: 0",axis=1)
valid_ind_data.head(5)

train_merged_data = pd.merge(train_ind_data,train_mkt_data, on ="Dates")
train_merged_data.head(5)
valid_merged_data = pd.merge(valid_ind_data,valid_mkt_data, on ="Dates")
valid_merged_data.head(5)

df1= train_merged_data.dropna(axis=1)
df1.to_csv('train_X_merged_for_feature_generating.csv')

df2= valid_merged_data.dropna(axis=1)
df2.to_csv('valid_X_merged_for_feature_generating.csv')
