==> earlier code/6. make features.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm_notebook as tqdm
#from tsfresh.feature_extraction import ComprehensiveFCParameters
#from tsfresh.feature_extraction import extract_features
#from tsfresh.utilities.dataframe_functions import roll_time_series
#pd.set_option('display.max_columns', 150)

!!!WARNING!!!
YOU MUST DEVIDE THE FEATURES BACK INTO TRAIN DATASET AND TEST DATASET!!!
the row index (not the row number) of the last row of train data is 4631

x_train = pd.read_csv("train_X_merged_for_feature_generating.csv",header = 0)
x_train= x_train.drop("Unnamed: 0",axis=1)
train_col = x_train.columns
x_valid = pd.read_csv("valid_X_merged_for_feature_generating.csv",header = 0)
x_valid=x_valid.drop("Unnamed: 0",axis=1)
valid_col = x_valid.columns
x_valid

yx_train=pd.read_csv('train_yX.csv')
yx_train
y_valid=pd.read_excel('valid_y.xlsx')
y_valid

x = pd.concat([x_train,x_valid]).reset_index(drop=True) 
x=x.dropna(axis=1)
x.to_csv("train&Valid_X_merged_for_feature_generating.csv")
x

df=x.drop('Dates',axis=1)
df["column_id"] = 0
span = 10

features_rec=pd.DataFrame()

res_df_max=pd.DataFrame()
res_df_min=pd.DataFrame()
res_df_mean=pd.DataFrame()
res_df_median=pd.DataFrame()
res_df_std=pd.DataFrame()
res_df_var=pd.DataFrame()
res_df_lcd005=pd.DataFrame()
res_df_lcd01=pd.DataFrame()
res_df_lcd02=pd.DataFrame()

for i in range(span-1,len(df)):
    df_cut=df[i-span+1:i+1]
    #calculate the 6 features firstly.
    temp_df_max=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_min=pd.DataFrame(df_cut.min().astype(float)).T
    temp_df_mean=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_median=pd.DataFrame(df_cut.median().astype(float)).T
    temp_df_std=pd.DataFrame(df_cut.std().astype(float)).T
    temp_df_var=pd.DataFrame(df_cut.var().astype(float)).T
    # record the features
    res_df_max=pd.concat([res_df_max,temp_df_max],axis=0)
    res_df_min=pd.concat([res_df_min,temp_df_min],axis=0)
    res_df_mean=pd.concat([res_df_mean,temp_df_mean],axis=0)
    res_df_median=pd.concat([res_df_median,temp_df_median],axis=0)
    res_df_std=pd.concat([res_df_std,temp_df_std],axis=0)
    res_df_var=pd.concat([res_df_var,temp_df_var],axis=0)
    #calculate the large standard deviation where r=0.05,0.1 and 0.2, record the results
    temp_df_lcd005=pd.DataFrame(np.where(temp_df_std>0.05*(temp_df_max-temp_df_min),1,0))
    res_df_lcd005=pd.concat([res_df_lcd005,temp_df_lcd005],axis=0)
    temp_df_lcd01=pd.DataFrame(np.where(temp_df_std>0.1*(temp_df_max-temp_df_min),1,0))
    res_df_lcd01=pd.concat([res_df_lcd01,temp_df_lcd01],axis=0)
    temp_df_lcd02=pd.DataFrame(np.where(temp_df_std>0.2*(temp_df_max-temp_df_min),1,0))
    res_df_lcd02=pd.concat([res_df_lcd02,temp_df_lcd02],axis=0)

max_col_index = [col + f'_max_{span}' for col in df.columns]    
min_col_index = [col + f'_min_{span}' for col in df.columns]
mean_col_index = [col + f'_mean_{span}' for col in df.columns]
median_col_index = [col +f'_median_{span}' for col in df.columns]
std_col_index = [col + f'_std_{span}' for col in df.columns]
var_col_index = [col + f'_var_{span}' for col in df.columns]
lsd005_col_index = [col + f'_large_standard_deviation_'+str(0.05)+'_{span}' for col in df.columns]
lsd01_col_index = [col + f'_large_standard_deviation_'+str(0.1)+'_{span}' for col in df.columns]
lsd02_col_index = [col +f'_large_standard_deviation_'+str(0.2)+'_{span}' for col in df.columns]



res_df_max.columns=max_col_index 
res_df_min.columns=min_col_index
res_df_mean.columns=mean_col_index
res_df_median.columns=median_col_index
res_df_std.columns=std_col_index
res_df_var.columns=var_col_index
res_df_lcd005.columns=lsd005_col_index
res_df_lcd01.columns=lsd01_col_index
res_df_lcd02.columns=lsd02_col_index


for i in range(res_df_max.shape[1]):
    temp_df = pd.concat([res_df_max.iloc[:,i], res_df_min.iloc[:,i], res_df_mean.iloc[:,i],res_df_median.iloc[:,i], res_df_std.iloc[:,i], res_df_var.iloc[:,i],res_df_lcd005.iloc[:,i], res_df_lcd01.iloc[:,i], res_df_lcd02.iloc[:,i]], axis=1)
    features_rec=pd.concat([features_rec, temp_df], axis=1)



features_rec.index=range(span-1,span-1+len(features_rec))
features_rec.to_csv(f'features_{span}.csv',index=True)

df=x.drop('Dates',axis=1)
df["column_id"] = 0
span = 5

features_5=pd.DataFrame()

res_df_max=pd.DataFrame()
res_df_min=pd.DataFrame()
res_df_mean=pd.DataFrame()
res_df_median=pd.DataFrame()
res_df_std=pd.DataFrame()
res_df_var=pd.DataFrame()
res_df_lcd005=pd.DataFrame()
res_df_lcd01=pd.DataFrame()
res_df_lcd02=pd.DataFrame()

for i in range(span-1,len(df)):
    df_cut=df[i-span+1:i+1]
    #calculate the 6 features firstly.
    temp_df_max=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_min=pd.DataFrame(df_cut.min().astype(float)).T
    temp_df_mean=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_median=pd.DataFrame(df_cut.median().astype(float)).T
    temp_df_std=pd.DataFrame(df_cut.std().astype(float)).T
    temp_df_var=pd.DataFrame(df_cut.var().astype(float)).T
    # record the features
    res_df_max=pd.concat([res_df_max,temp_df_max],axis=0)
    res_df_min=pd.concat([res_df_min,temp_df_min],axis=0)
    res_df_mean=pd.concat([res_df_mean,temp_df_mean],axis=0)
    res_df_median=pd.concat([res_df_median,temp_df_median],axis=0)
    res_df_std=pd.concat([res_df_std,temp_df_std],axis=0)
    res_df_var=pd.concat([res_df_var,temp_df_var],axis=0)
    #calculate the large standard deviation where r=0.05,0.1 and 0.2, record the results
    temp_df_lcd005=pd.DataFrame(np.where(temp_df_std>0.05*(temp_df_max-temp_df_min),1,0))
    res_df_lcd005=pd.concat([res_df_lcd005,temp_df_lcd005],axis=0)
    temp_df_lcd01=pd.DataFrame(np.where(temp_df_std>0.1*(temp_df_max-temp_df_min),1,0))
    res_df_lcd01=pd.concat([res_df_lcd01,temp_df_lcd01],axis=0)
    temp_df_lcd02=pd.DataFrame(np.where(temp_df_std>0.2*(temp_df_max-temp_df_min),1,0))
    res_df_lcd02=pd.concat([res_df_lcd02,temp_df_lcd02],axis=0)

max_col_index = [col + '_max_5' for col in df.columns]    
min_col_index = [col + '_min_5' for col in df.columns]
mean_col_index = [col + '_mean_5' for col in df.columns]
median_col_index = [col + '_median_5' for col in df.columns]
std_col_index = [col + '_std_5' for col in df.columns]
var_col_index = [col + '_var_5' for col in df.columns]
lsd005_col_index = [col + '_large_standard_deviation_'+str(0.05)+'_5' for col in df.columns]
lsd01_col_index = [col + '_large_standard_deviation_'+str(0.1)+'_5' for col in df.columns]
lsd02_col_index = [col +'_large_standard_deviation_'+str(0.2)+'_5' for col in df.columns]



res_df_max.columns=max_col_index 
res_df_min.columns=min_col_index
res_df_mean.columns=mean_col_index
res_df_median.columns=median_col_index
res_df_std.columns=std_col_index
res_df_var.columns=var_col_index
res_df_lcd005.columns=lsd005_col_index
res_df_lcd01.columns=lsd01_col_index
res_df_lcd02.columns=lsd02_col_index
'''res_df_max=res_df_max.rename(columns=lambda x: x + '_max_5')
res_df_min=res_df_min.rename(columns=lambda x: x + '_min_5')
res_df_mean=res_df_mean.rename(columns=lambda x: x + '_mean_5')
res_df_median=res_df_median.rename(columns=lambda x: x + '_median_5')
res_df_std=res_df_std.rename(columns=lambda x: x + '_std_5')
res_df_var=res_df_var.rename(columns=lambda x: x + '_var_5')
res_df_lcd005=res_df_lcd005.rename(columns=lambda x: str(x) + '_large_standard_deviation_'+str(0.05)+'_5')
res_df_lcd01=res_df_lcd01.rename(columns=lambda x: str(x) + '_large_standard_deviation_'+str(0.1)+'_5')
res_df_lcd02=res_df_lcd02.rename(columns=lambda x: str(x)+ '_large_standard_deviation_'+str(0.2)+'_5')'''


for i in range(res_df_max.shape[1]):
    temp_df = pd.concat([res_df_max.iloc[:,i], res_df_min.iloc[:,i], res_df_mean.iloc[:,i],res_df_median.iloc[:,i], res_df_std.iloc[:,i], res_df_var.iloc[:,i],res_df_lcd005.iloc[:,i], res_df_lcd01.iloc[:,i], res_df_lcd02.iloc[:,i]], axis=1)
    features_5=pd.concat([features_5, temp_df], axis=1)

features_5.index=range(span-1,span-1+len(features_5))
features_5.to_csv('features_5.csv',index=True)

df=x.drop('Dates',axis=1)
df["column_id"] = 0
span = 10

features_10=pd.DataFrame()

res_df_max=pd.DataFrame()
res_df_min=pd.DataFrame()
res_df_mean=pd.DataFrame()
res_df_median=pd.DataFrame()
res_df_std=pd.DataFrame()
res_df_var=pd.DataFrame()
res_df_lcd005=pd.DataFrame()
res_df_lcd01=pd.DataFrame()
res_df_lcd02=pd.DataFrame()

for i in range(span-1,len(df)):
    df_cut=df[i-span+1:i+1]
    #calculate the 6 features firstly.
    temp_df_max=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_min=pd.DataFrame(df_cut.min().astype(float)).T
    temp_df_mean=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_median=pd.DataFrame(df_cut.median().astype(float)).T
    temp_df_std=pd.DataFrame(df_cut.std().astype(float)).T
    temp_df_var=pd.DataFrame(df_cut.var().astype(float)).T
    # record the features
    res_df_max=pd.concat([res_df_max,temp_df_max],axis=0)
    res_df_min=pd.concat([res_df_min,temp_df_min],axis=0)
    res_df_mean=pd.concat([res_df_mean,temp_df_mean],axis=0)
    res_df_median=pd.concat([res_df_median,temp_df_median],axis=0)
    res_df_std=pd.concat([res_df_std,temp_df_std],axis=0)
    res_df_var=pd.concat([res_df_var,temp_df_var],axis=0)
    #calculate the large standard deviation where r=0.05,0.1 and 0.2, record the results
    temp_df_lcd005=pd.DataFrame(np.where(temp_df_std>0.05*(temp_df_max-temp_df_min),1,0))
    res_df_lcd005=pd.concat([res_df_lcd005,temp_df_lcd005],axis=0)
    temp_df_lcd01=pd.DataFrame(np.where(temp_df_std>0.1*(temp_df_max-temp_df_min),1,0))
    res_df_lcd01=pd.concat([res_df_lcd01,temp_df_lcd01],axis=0)
    temp_df_lcd02=pd.DataFrame(np.where(temp_df_std>0.2*(temp_df_max-temp_df_min),1,0))
    res_df_lcd02=pd.concat([res_df_lcd02,temp_df_lcd02],axis=0)

max_col_index = [col + '_max_10' for col in df.columns]    
min_col_index = [col + '_min_10' for col in df.columns]
mean_col_index = [col + '_mean_10' for col in df.columns]
median_col_index = [col + '_median_10' for col in df.columns]
std_col_index = [col + '_std_10' for col in df.columns]
var_col_index = [col + '_var_10' for col in df.columns]
lsd005_col_index = [col + '_large_standard_deviation_'+str(0.05)+'_10' for col in df.columns]
lsd01_col_index = [col + '_large_standard_deviation_'+str(0.1)+'_10' for col in df.columns]
lsd02_col_index = [col +'_large_standard_deviation_'+str(0.2)+'_10' for col in df.columns]



res_df_max.columns=max_col_index 
res_df_min.columns=min_col_index
res_df_mean.columns=mean_col_index
res_df_median.columns=median_col_index
res_df_std.columns=std_col_index
res_df_var.columns=var_col_index
res_df_lcd005.columns=lsd005_col_index
res_df_lcd01.columns=lsd01_col_index
res_df_lcd02.columns=lsd02_col_index


for i in range(res_df_max.shape[1]):
    temp_df = pd.concat([res_df_max.iloc[:,i], res_df_min.iloc[:,i], res_df_mean.iloc[:,i],res_df_median.iloc[:,i], res_df_std.iloc[:,i], res_df_var.iloc[:,i],res_df_lcd005.iloc[:,i], res_df_lcd01.iloc[:,i], res_df_lcd02.iloc[:,i]], axis=1)
    features_10=pd.concat([features_10, temp_df], axis=1)

features_10.index=range(span-1,span-1+len(features_10))
features_10.to_csv('features_10.csv',index=True)

df=x.drop('Dates',axis=1)
df["column_id"] = 0
span = 25

features_25=pd.DataFrame()

res_df_max=pd.DataFrame()
res_df_min=pd.DataFrame()
res_df_mean=pd.DataFrame()
res_df_median=pd.DataFrame()
res_df_std=pd.DataFrame()
res_df_var=pd.DataFrame()
res_df_lcd005=pd.DataFrame()
res_df_lcd01=pd.DataFrame()
res_df_lcd02=pd.DataFrame()

for i in range(span-1,len(df)):
    df_cut=df[i-span+1:i+1]
    #calculate the 6 features firstly.
    temp_df_max=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_min=pd.DataFrame(df_cut.min().astype(float)).T
    temp_df_mean=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_median=pd.DataFrame(df_cut.median().astype(float)).T
    temp_df_std=pd.DataFrame(df_cut.std().astype(float)).T
    temp_df_var=pd.DataFrame(df_cut.var().astype(float)).T
    # record the features
    res_df_max=pd.concat([res_df_max,temp_df_max],axis=0)
    res_df_min=pd.concat([res_df_min,temp_df_min],axis=0)
    res_df_mean=pd.concat([res_df_mean,temp_df_mean],axis=0)
    res_df_median=pd.concat([res_df_median,temp_df_median],axis=0)
    res_df_std=pd.concat([res_df_std,temp_df_std],axis=0)
    res_df_var=pd.concat([res_df_var,temp_df_var],axis=0)
    #calculate the large standard deviation where r=0.05,0.1 and 0.2, record the results
    temp_df_lcd005=pd.DataFrame(np.where(temp_df_std>0.05*(temp_df_max-temp_df_min),1,0))
    res_df_lcd005=pd.concat([res_df_lcd005,temp_df_lcd005],axis=0)
    temp_df_lcd01=pd.DataFrame(np.where(temp_df_std>0.1*(temp_df_max-temp_df_min),1,0))
    res_df_lcd01=pd.concat([res_df_lcd01,temp_df_lcd01],axis=0)
    temp_df_lcd02=pd.DataFrame(np.where(temp_df_std>0.2*(temp_df_max-temp_df_min),1,0))
    res_df_lcd02=pd.concat([res_df_lcd02,temp_df_lcd02],axis=0)

max_col_index = [col + '_max_25' for col in df.columns]    
min_col_index = [col + '_min_25' for col in df.columns]
mean_col_index = [col + '_mean_25' for col in df.columns]
median_col_index = [col + '_median_25' for col in df.columns]
std_col_index = [col + '_std_25' for col in df.columns]
var_col_index = [col + '_var_25' for col in df.columns]
lsd005_col_index = [col + '_large_standard_deviation_'+str(0.05)+'_25' for col in df.columns]
lsd01_col_index = [col + '_large_standard_deviation_'+str(0.1)+'_25' for col in df.columns]
lsd02_col_index = [col +'_large_standard_deviation_'+str(0.2)+'_25' for col in df.columns]



res_df_max.columns=max_col_index 
res_df_min.columns=min_col_index
res_df_mean.columns=mean_col_index
res_df_median.columns=median_col_index
res_df_std.columns=std_col_index
res_df_var.columns=var_col_index
res_df_lcd005.columns=lsd005_col_index
res_df_lcd01.columns=lsd01_col_index
res_df_lcd02.columns=lsd02_col_index


for i in range(res_df_max.shape[1]):
    temp_df = pd.concat([res_df_max.iloc[:,i], res_df_min.iloc[:,i], res_df_mean.iloc[:,i],res_df_median.iloc[:,i], res_df_std.iloc[:,i], res_df_var.iloc[:,i],res_df_lcd005.iloc[:,i], res_df_lcd01.iloc[:,i], res_df_lcd02.iloc[:,i]], axis=1)
    features_25=pd.concat([features_25, temp_df], axis=1)

features_25.index=range(span-1,span-1+len(features_25))
features_25.to_csv('features_25.csv',index=True)

df=x.drop('Dates',axis=1)
df["column_id"] = 0
span = 50

features_50=pd.DataFrame()

res_df_max=pd.DataFrame()
res_df_min=pd.DataFrame()
res_df_mean=pd.DataFrame()
res_df_median=pd.DataFrame()
res_df_std=pd.DataFrame()
res_df_var=pd.DataFrame()
res_df_lcd005=pd.DataFrame()
res_df_lcd01=pd.DataFrame()
res_df_lcd02=pd.DataFrame()

for i in range(span-1,len(df)):
    df_cut=df[i-span+1:i+1]
    #calculate the 6 features firstly.
    temp_df_max=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_min=pd.DataFrame(df_cut.min().astype(float)).T
    temp_df_mean=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_median=pd.DataFrame(df_cut.median().astype(float)).T
    temp_df_std=pd.DataFrame(df_cut.std().astype(float)).T
    temp_df_var=pd.DataFrame(df_cut.var().astype(float)).T
    # record the features
    res_df_max=pd.concat([res_df_max,temp_df_max],axis=0)
    res_df_min=pd.concat([res_df_min,temp_df_min],axis=0)
    res_df_mean=pd.concat([res_df_mean,temp_df_mean],axis=0)
    res_df_median=pd.concat([res_df_median,temp_df_median],axis=0)
    res_df_std=pd.concat([res_df_std,temp_df_std],axis=0)
    res_df_var=pd.concat([res_df_var,temp_df_var],axis=0)
    #calculate the large standard deviation where r=0.05,0.1 and 0.2, record the results
    temp_df_lcd005=pd.DataFrame(np.where(temp_df_std>0.05*(temp_df_max-temp_df_min),1,0))
    res_df_lcd005=pd.concat([res_df_lcd005,temp_df_lcd005],axis=0)
    temp_df_lcd01=pd.DataFrame(np.where(temp_df_std>0.1*(temp_df_max-temp_df_min),1,0))
    res_df_lcd01=pd.concat([res_df_lcd01,temp_df_lcd01],axis=0)
    temp_df_lcd02=pd.DataFrame(np.where(temp_df_std>0.2*(temp_df_max-temp_df_min),1,0))
    res_df_lcd02=pd.concat([res_df_lcd02,temp_df_lcd02],axis=0)

max_col_index = [col + '_max_50' for col in df.columns]    
min_col_index = [col + '_min_50' for col in df.columns]
mean_col_index = [col + '_mean_50' for col in df.columns]
median_col_index = [col + '_median_50' for col in df.columns]
std_col_index = [col + '_std_50' for col in df.columns]
var_col_index = [col + '_var_50' for col in df.columns]
lsd005_col_index = [col + '_large_standard_deviation_'+str(0.05)+'_50' for col in df.columns]
lsd01_col_index = [col + '_large_standard_deviation_'+str(0.1)+'_50' for col in df.columns]
lsd02_col_index = [col +'_large_standard_deviation_'+str(0.2)+'_50' for col in df.columns]



res_df_max.columns=max_col_index 
res_df_min.columns=min_col_index
res_df_mean.columns=mean_col_index
res_df_median.columns=median_col_index
res_df_std.columns=std_col_index
res_df_var.columns=var_col_index
res_df_lcd005.columns=lsd005_col_index
res_df_lcd01.columns=lsd01_col_index
res_df_lcd02.columns=lsd02_col_index


for i in range(res_df_max.shape[1]):
    temp_df = pd.concat([res_df_max.iloc[:,i], res_df_min.iloc[:,i], res_df_mean.iloc[:,i],res_df_median.iloc[:,i], res_df_std.iloc[:,i], res_df_var.iloc[:,i],res_df_lcd005.iloc[:,i], res_df_lcd01.iloc[:,i], res_df_lcd02.iloc[:,i]], axis=1)
    features_50=pd.concat([features_50, temp_df], axis=1)

features_50.index=range(span-1,span-1+len(features_50))
features_50.to_csv('features_50.csv',index=True)

df=x.drop('Dates',axis=1)
df["column_id"] = 0
span = 100

features_100=pd.DataFrame()

res_df_max=pd.DataFrame()
res_df_min=pd.DataFrame()
res_df_mean=pd.DataFrame()
res_df_median=pd.DataFrame()
res_df_std=pd.DataFrame()
res_df_var=pd.DataFrame()
res_df_lcd005=pd.DataFrame()
res_df_lcd01=pd.DataFrame()
res_df_lcd02=pd.DataFrame()

for i in range(span-1,len(df)):
    df_cut=df[i-span+1:i+1]
    #calculate the 6 features firstly.
    temp_df_max=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_min=pd.DataFrame(df_cut.min().astype(float)).T
    temp_df_mean=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_median=pd.DataFrame(df_cut.median().astype(float)).T
    temp_df_std=pd.DataFrame(df_cut.std().astype(float)).T
    temp_df_var=pd.DataFrame(df_cut.var().astype(float)).T
    # record the features
    res_df_max=pd.concat([res_df_max,temp_df_max],axis=0)
    res_df_min=pd.concat([res_df_min,temp_df_min],axis=0)
    res_df_mean=pd.concat([res_df_mean,temp_df_mean],axis=0)
    res_df_median=pd.concat([res_df_median,temp_df_median],axis=0)
    res_df_std=pd.concat([res_df_std,temp_df_std],axis=0)
    res_df_var=pd.concat([res_df_var,temp_df_var],axis=0)
    #calculate the large standard deviation where r=0.05,0.1 and 0.2, record the results
    temp_df_lcd005=pd.DataFrame(np.where(temp_df_std>0.05*(temp_df_max-temp_df_min),1,0))
    res_df_lcd005=pd.concat([res_df_lcd005,temp_df_lcd005],axis=0)
    temp_df_lcd01=pd.DataFrame(np.where(temp_df_std>0.1*(temp_df_max-temp_df_min),1,0))
    res_df_lcd01=pd.concat([res_df_lcd01,temp_df_lcd01],axis=0)
    temp_df_lcd02=pd.DataFrame(np.where(temp_df_std>0.2*(temp_df_max-temp_df_min),1,0))
    res_df_lcd02=pd.concat([res_df_lcd02,temp_df_lcd02],axis=0)

max_col_index = [col + '_max_100' for col in df.columns]    
min_col_index = [col + '_min_100' for col in df.columns]
mean_col_index = [col + '_mean_100' for col in df.columns]
median_col_index = [col + '_median_100' for col in df.columns]
std_col_index = [col + '_std_100' for col in df.columns]
var_col_index = [col + '_var_100' for col in df.columns]
lsd005_col_index = [col + '_large_standard_deviation_'+str(0.05)+'_100' for col in df.columns]
lsd01_col_index = [col + '_large_standard_deviation_'+str(0.1)+'_100' for col in df.columns]
lsd02_col_index = [col +'_large_standard_deviation_'+str(0.2)+'_100' for col in df.columns]



res_df_max.columns=max_col_index 
res_df_min.columns=min_col_index
res_df_mean.columns=mean_col_index
res_df_median.columns=median_col_index
res_df_std.columns=std_col_index
res_df_var.columns=var_col_index
res_df_lcd005.columns=lsd005_col_index
res_df_lcd01.columns=lsd01_col_index
res_df_lcd02.columns=lsd02_col_index


for i in range(res_df_max.shape[1]):
    temp_df = pd.concat([res_df_max.iloc[:,i], res_df_min.iloc[:,i], res_df_mean.iloc[:,i],res_df_median.iloc[:,i], res_df_std.iloc[:,i], res_df_var.iloc[:,i],res_df_lcd005.iloc[:,i], res_df_lcd01.iloc[:,i], res_df_lcd02.iloc[:,i]], axis=1)
    features_100=pd.concat([features_100, temp_df], axis=1)

features_100.index=range(span-1,span-1+len(features_100))
features_100.to_csv('features_100.csv',index=True)

df=x.drop('Dates',axis=1)
df["column_id"] = 0
span = 200

features_200=pd.DataFrame()

res_df_max=pd.DataFrame()
res_df_min=pd.DataFrame()
res_df_mean=pd.DataFrame()
res_df_median=pd.DataFrame()
res_df_std=pd.DataFrame()
res_df_var=pd.DataFrame()
res_df_lcd005=pd.DataFrame()
res_df_lcd01=pd.DataFrame()
res_df_lcd02=pd.DataFrame()

for i in range(span-1,len(df)):
    df_cut=df[i-span+1:i+1]
    #calculate the 6 features firstly.
    temp_df_max=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_min=pd.DataFrame(df_cut.min().astype(float)).T
    temp_df_mean=pd.DataFrame(df_cut.mean().astype(float)).T
    temp_df_median=pd.DataFrame(df_cut.median().astype(float)).T
    temp_df_std=pd.DataFrame(df_cut.std().astype(float)).T
    temp_df_var=pd.DataFrame(df_cut.var().astype(float)).T
    # record the features
    res_df_max=pd.concat([res_df_max,temp_df_max],axis=0)
    res_df_min=pd.concat([res_df_min,temp_df_min],axis=0)
    res_df_mean=pd.concat([res_df_mean,temp_df_mean],axis=0)
    res_df_median=pd.concat([res_df_median,temp_df_median],axis=0)
    res_df_std=pd.concat([res_df_std,temp_df_std],axis=0)
    res_df_var=pd.concat([res_df_var,temp_df_var],axis=0)
    #calculate the large standard deviation where r=0.05,0.1 and 0.2, record the results
    temp_df_lcd005=pd.DataFrame(np.where(temp_df_std>0.05*(temp_df_max-temp_df_min),1,0))
    res_df_lcd005=pd.concat([res_df_lcd005,temp_df_lcd005],axis=0)
    temp_df_lcd01=pd.DataFrame(np.where(temp_df_std>0.1*(temp_df_max-temp_df_min),1,0))
    res_df_lcd01=pd.concat([res_df_lcd01,temp_df_lcd01],axis=0)
    temp_df_lcd02=pd.DataFrame(np.where(temp_df_std>0.2*(temp_df_max-temp_df_min),1,0))
    res_df_lcd02=pd.concat([res_df_lcd02,temp_df_lcd02],axis=0)

max_col_index = [col + '_max_200' for col in df.columns]    
min_col_index = [col + '_min_200' for col in df.columns]
mean_col_index = [col + '_mean_200' for col in df.columns]
median_col_index = [col + '_median_200' for col in df.columns]
std_col_index = [col + '_std_200' for col in df.columns]
var_col_index = [col + '_var_200' for col in df.columns]
lsd005_col_index = [col + '_large_standard_deviation_'+str(0.05)+'_200' for col in df.columns]
lsd01_col_index = [col + '_large_standard_deviation_'+str(0.1)+'_200' for col in df.columns]
lsd02_col_index = [col +'_large_standard_deviation_'+str(0.2)+'_200' for col in df.columns]



res_df_max.columns=max_col_index 
res_df_min.columns=min_col_index
res_df_mean.columns=mean_col_index
res_df_median.columns=median_col_index
res_df_std.columns=std_col_index
res_df_var.columns=var_col_index
res_df_lcd005.columns=lsd005_col_index
res_df_lcd01.columns=lsd01_col_index
res_df_lcd02.columns=lsd02_col_index


for i in range(res_df_max.shape[1]):
    temp_df = pd.concat([res_df_max.iloc[:,i], res_df_min.iloc[:,i], res_df_mean.iloc[:,i],res_df_median.iloc[:,i], res_df_std.iloc[:,i], res_df_var.iloc[:,i],res_df_lcd005.iloc[:,i], res_df_lcd01.iloc[:,i], res_df_lcd02.iloc[:,i]], axis=1)
    features_200=pd.concat([features_200, temp_df], axis=1)

features_200.index=range(span-1,span-1+len(features_200))
features_200.to_csv('features_200.csv',index=True)

x.loc[4631,:]

features_200.index=range(span-1,span-1+len(features_200))
features_200.to_csv('features_200.csv',index=True)

features_200

x.loc[4631]
