==> earlier code/7. features divide into train and valid.ipynb <==
import pandas as pd

df5=pd.read_csv('features_5.csv')
df10=pd.read_csv('features_10.csv')
df25=pd.read_csv('features_25.csv')
df50=pd.read_csv('features_50.csv')
df100=pd.read_csv('features_100.csv')
df200=pd.read_csv('features_200.csv')

df5

df5.set_index(df5.iloc[:, 0], inplace=True)
df5=df5.drop("Unnamed: 0",axis=1)
df5_train=df5.loc[:4631]
df5_valid=df5.loc[4632:]
df5_train.to_csv('features_5_train.csv')
df5_valid.to_csv('features_5_valid.csv')

df10.set_index(df10.iloc[:, 0], inplace=True)
df10=df10.drop("Unnamed: 0",axis=1)
df10_train=df10.loc[:4631]
df10_valid=df10.loc[4632:]
df10_train.to_csv('features_10_train.csv')
df10_valid.to_csv('features_10_valid.csv')

df25.set_index(df25.iloc[:, 0], inplace=True)
df25=df25.drop("Unnamed: 0",axis=1)
df25_train=df25.loc[:4631]
df25_valid=df25.loc[4632:]
df25_train.to_csv('features_25_train.csv')
df25_valid.to_csv('features_25_valid.csv')

df50.set_index(df50.iloc[:, 0], inplace=True)
df50=df50.drop("Unnamed: 0",axis=1)
df50_train=df50.loc[:4631]
df50_valid=df50.loc[4632:]
df50_train.to_csv('features_50_train.csv')
df50_valid.to_csv('features_50_valid.csv')

df100.set_index(df100.iloc[:, 0], inplace=True)
df100=df100.drop("Unnamed: 0",axis=1)
df100_train=df100.loc[:4631]
df100_valid=df100.loc[4632:]
df100_train.to_csv('features_100_train.csv')
df100_valid.to_csv('features_100_valid.csv')

df200.set_index(df200.iloc[:, 0], inplace=True)
df200=df200.drop("Unnamed: 0",axis=1)
#df200.to_csv('df200featuretemp.csv')
df200_train=df200.loc[:4631]
df200_valid=df200.loc[4632:]
df200_train.to_csv('features_200_train.csv')
df200_valid.to_csv('features_200_valid.csv')

df000=pd.read_csv('features_100_valid.csv')
df000
