==> earlier code/8. make technical features.ipynb <==
import talib as ta
import pandas as pd
import numpy as np

df = pd.read_csv("train&Valid_X_merged_for_feature_generating.csv",header = 0)
df["Dates"] = pd.to_datetime(df["Dates"])
df

period = 10
sma_value = ta.SMA(np.array(df['TargetPortDailyRtn']), period)
ema_value = ta.EMA(np.array(df['TargetPortDailyRtn']), period)
fast_period = 8
slow_period = 18
signal_period = 6
macd, macd_signal, macd_hist = ta.MACD(np.array(df['TargetPortDailyRtn']), fast_period,
                                          slow_period, signal_period)
rsi_values = ta.RSI(np.array(df['TargetPortDailyRtn']), period)
sigma = 2
moving_average_type = 0
bb_up, bb_middle, bb_down = ta.BBANDS(np.array(df['TargetPortDailyRtn']),
                                         period, sigma, sigma, moving_average_type)
df["sma_value"] = sma_value
df["ema_value"] = ema_value
df["macd"] = macd
df["macd_signal"] = macd_signal
df["macd_hist"] = macd_hist
df["rsi_values"] = rsi_values
df["bb_up"] = bb_up
df["bb_middle"] = bb_middle
df["bb_down"] = bb_down

df = df[["Dates","sma_value","ema_value","macd","macd_signal","macd_hist","rsi_values","bb_up","bb_middle","bb_down"]]
df

new_columns = ['Dates_0']
for column in df.columns:
    if column != "Dates":
        new_columns.append(column+"_0")
df.columns = new_columns

df.to_csv('train&valid_technical_features.csv')

df
