==> earlier code/9. feratures further process.ipynb <==
!!!READ THIS!!!!!!

here , information of date and corresponding y data is added to the features, so that each row in the features includes the infomation of DATE,DEC_RATE,SIGNAL, and the features from training set

import numpy 
import pandas as pd

df=pd.read_csv('train&valid_yX_after_process.csv')
df_info=df.iloc[:,:11]
df_info=df_info.drop(df_info.columns[[0,1]],axis=1)
df_info

span=200
df_features=pd.read_csv(f'features_{span}.csv')
df_features.index=df_features.iloc[:,0]
df_features_new=pd.concat([df_info,df_features],axis=1)
df_features_new=df_features_new.dropna(axis=0)
df_features_new.to_csv(f'features_new_{span}.csv')
