==> earlier code/10. NN run01.ipynb <==
import os

import pandas as pd
import numpy as np
import torch
import torch.nn as nn
import torch.nn.functional as F
import torch.optim as optim
from torch.autograd import Variable 
from torch.utils.data import DataLoader
from torch.utils.data import TensorDataset


from sklearn.metrics import roc_curve,roc_auc_score
import matplotlib.pyplot as plt

device = 'cuda' if torch.cuda.is_available() else 'cpu'
device

特徴量を読み込み、学習の前に必要な前処理を行う
1．正規化
2．train、validデータに分け、そして内部もXとｙに分ける
3．NN入力の形に変える

from sklearn.preprocessing import MinMaxScaler

do noramalization to the columns form col index10

scaler = MinMaxScaler()
span=50
df=pd.read_csv(f'features_new_{span}.csv')
df=df.drop(df.columns[[10]],axis=1)#drop the meaningless col, which is the row index
df.index=df.iloc[:,0]
df=df.drop(['Dates'],axis=1)
df.iloc[:, 10:] = scaler.fit_transform(df.iloc[:, 10:])
df=df.drop(df.columns[[0]],axis=1)
df

split the features data into trainX/y and validX/y

def split_data(features,batch_size):
    features_y=features.iloc[:,8]
    features_x=features.drop(features.columns[[8,9]],axis=1)
    y_raw=features_y.to_numpy()
    x_raw=features_x.to_numpy()
    data_x=[]
    data_y=[]
    for i in range(len(x_raw)-batch_size):
        data_x.append(x_raw[i:i+batch_size])
        data_y.append(y_raw[i:i+batch_size])
    split_position=4631-span-batch_size+3
    data_x=np.array(data_x)
    data_y=np.array(data_y)
    x_train=data_x[:split_position]
    x_test=data_x[split_position:]
    y_train=data_y[:split_position]
    y_test=data_y[split_position:]
    return [x_train, y_train, x_test, y_test]

batch_size=10
x_train, y_train, x_test, y_test = split_data(df, batch_size)

x_train=x_train.astype(np.float64)
y_train=y_train.astype(np.float64)
x_test=x_test.astype(np.float64)
y_test=y_test.astype(np.float64)

print('x_train.shape = ',x_train.shape)
print('y_train.shape = ',y_train.shape)
print('x_test.shape = ',x_test.shape)
print('y_test.shape = ',y_test.shape)

x_test

処理した特徴量をNNに入力するtensorの形に変える

x_train = torch.from_numpy(x_train).type(torch.Tensor)
x_test = torch.from_numpy(x_test).type(torch.Tensor)
y_train_lstm = torch.from_numpy(y_train).type(torch.Tensor)
y_test_lstm = torch.from_numpy(y_test).type(torch.Tensor)
y_train_gru = torch.from_numpy(y_train).type(torch.Tensor)
y_test_gru = torch.from_numpy(y_test).type(torch.Tensor)

NNの構造を決める

n_features = 744
hidden_size = 40
output_size = 2
width = 5
epochs = 100
batch_size = 10
test_batch = 1
num_layers = 2

lr = 0.01
save=False
name = 'model_weights1.pth'

class LSTM_model(nn.Module):
    def __init__(self,n_features,width,hidden_size,output_size,num_layers):
        super(LSTM_model,self).__init__()
        self.lstm = nn.LSTM(n_features, hidden_size, num_layers=num_layers)
        self.fc = nn.Linear(hidden_size, output_size)
        self.softmax = nn.Softmax(dim=1)

    def forward(self,x):
        lstm_out, _ = self.lstm(x)
        lstm_out = lstm_out[:, -1, :]
        out = self.fc(lstm_out)
        out = self.softmax(out)
        return out

model = LSTM_model(n_features=n_features,width=width,hidden_size=hidden_size,output_size=output_size,num_layers=num_layers).to(device)
criterion = torch.nn.MSELoss(reduction='mean')
optimiser = torch.optim.Adam(model.parameters(), lr=0.01)

x_train[0].shape

train_data = TensorDataset(x_train, y_train_lstm)
test_data = TensorDataset(x_test, y_test_lstm)
train_loader = DataLoader(train_data, batch_size=batch_size)
test_loader=DataLoader(test_data,batch_size=batch_size)

train_losses = []
test_losses = []

early_stopping_counter=0
patience=10

for epoch in range(epochs):
    model.train()
    train_loss=0
    val_loss=0
    for x_batch, y_batch in train_loader:
        y_pred = model(x_batch)
        loss = criterion(y_pred, y_batch)
        
        optimiser.zero_grad()
        loss.backward()
        optimiser.step()
        train_loss += loss.item() * x_batch.size(0)
    train_loss /= len(train_loader.dataset)
    
    

predict = pd.DataFrame(scaler.inverse_transform(y_train_pred.detach().numpy()))
original = pd.DataFrame(scaler.inverse_transform(y_train_lstm.detach().numpy()))
